// src/app/shared/model/Order.ts
/**
 * STEAK
 * Kantinen-API
 *
 * OpenAPI spec version: 0.0.2
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import * as models from './models';

export interface Order {
    /**
     * The ID given by the database
     */
    _id?: string;

    /**
     * ...
     */
    employee_id: string;

    offer?: models.Offer;

    /**
     * Is takeaway
     */
    takeaway_flag: boolean;

    /**
     * ...
     */
    date?: Date | string;

    /**
     * The price paid by the employee. some pay less because of their status.
     */
    amount: number;

    /**
     * ...
     */
    paid: boolean;

}